refactor: nearby
