+++ app/src/main/java/com/se2/bopit/domain/models/NearbyPayload.java
@@ -1,0 +1,13 @@
+package com.se2.bopit.domain.models;
+
+import java.io.Serializable;
+
+public class NearbyPayload implements Serializable{
+    public String payload;
+    public int type;
+
+    public NearbyPayload(int type, String payload){
+        this.type = type;
+        this.payload = payload;
+    }
+}
